refactor ILocation to fix user model functions import
